ui(refactor): split 'superdesk-content-filters' module.
