+++ scripts/superdesk-content-filters/controllers/index.js
@@ -1,0 +1,5 @@
+export { ContentFiltersConfigController } from './ContentFiltersController';
+export { FilterConditionsController } from './FilterConditionsController';
+export { ManageContentFiltersController } from './ManageContentFiltersController';
+export { ProductionTestController } from './ProductionTestController';
+export { FilterSearchController } from './FilterSearchController';
