+++ scripts/superdesk-content-filters/directives/index.js
@@ -1,0 +1,1 @@
+export { ManageFiltersTab } from './ManageFiltersTab';
