+++ scripts/superdesk-content-filters/services/index.js
@@ -1,0 +1,1 @@
+export { ContentFiltersService } from './ContentFiltersService';
